% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R
\docType{data}
\name{basketball}
\alias{basketball}
\title{College basketball data}
\format{
A data frame with 363 rows and 18 columns:
\describe{
\item{School}{School (character)}
\item{State}{State (character)}
\item{W}{Wins (integer)}
\item{L}{Loss's (integer)}
\item{W.L}{Win Loss percentage (numeric)}
\item{SRS}{Simple Rating System (numeric)}
\item{SOS}{Strength of Schedule (numeric)}
\item{Points.Scored}{Points scored (integer)}
\item{Points.Allowed}{Points allowed (integer)}
\item{FG.}{Team field goal percentage (numeric)}
\item{X3P.}{Three point percentage (numeric)}
\item{FT.}{Free throw percentage (numeric)}
\item{Rebounds}{Number of rebounds (integer)}
\item{AST}{Number of assists (integer)}
\item{STL}{Number of steals (integer)}
\item{Blocks}{Number of blocks (integer)}
\item{Turn.Overs}{Number of turn overs (integer)}
\item{Fouls}{Number of fouls (integer)}
}
}
\source{
Data retrieved from
\href{https://www.sports-reference.com/cbb/seasons/men/2023-school-stats.html}{Sports Reference}
with alterations made for educational purposes.
}
\usage{
basketball
}
\description{
This dataset contains performance statistics for 363 men’s college
basketball teams from the 2022-23 season.
}
\keyword{datasets}
