% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_demographics.R
\docType{data}
\name{class_demographics}
\alias{class_demographics}
\title{Class demographics
A sample dataset representing demographic and academic information for 50
college students.}
\format{
A data frame with 50 rows and 6 columns:
\describe{
\item{name}{Persons name (character)}
\item{ages}{Persons age (int)}
\item{state}{Persons state (character)}
\item{year}{Persons year in college (character)}
\item{major}{Persons major (character)}
\item{sport}{Binary Sport, 1(yes) or 0(no) (integer)}
}
}
\source{
Synthetic Data
}
\usage{
class_demographics
}
\description{
Class demographics
A sample dataset representing demographic and academic information for 50
college students.
}
\keyword{datasets}
