% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/election_2020.R
\docType{data}
\name{election_2020}
\alias{election_2020}
\title{2020 election data}
\format{
A data frame with 32177 rows and 7 columns:
\describe{
\item{state}{State (character)}
\item{state_ev}{State electoral votes (integer)}
\item{county}{County name (character)}
\item{candidate}{Candidate name (character)}
\item{party}{Candidate party (character)}
\item{total_votes}{Total number of votes (integer)}
\item{won}{True or false for the candidate to win the county (logical)}
}
}
\source{
2020 election data
}
\usage{
election_2020
}
\description{
2020 election data
}
\keyword{datasets}
