% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_scores.R
\docType{data}
\name{exam_scores}
\alias{exam_scores}
\title{Exam data}
\format{
A data frame with 99 rows and 8 columns:
\describe{
\item{gender}{Students gender (character)}
\item{race.ethnicity}{Students race/ethnicity (character)}
\item{parental.level.of.education}{Parents level of education (character)}
\item{lunch}{Students lunch plan (character)}
\item{test.preparation.course}{Student test prep level (character)}
\item{math.score}{Students math score (integer)}
\item{reading.score}{Students reading score (integer)}
\item{writing.score}{Students writing score (integer)}
}
}
\source{
Exam data
}
\usage{
exam_scores
}
\description{
Exam data
}
\keyword{datasets}
