% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basketball.R, R/football.R
\docType{data}
\name{football}
\alias{football}
\title{Ncaa basketball data}
\format{
A data frame with 363 rows and 18 columns:
\describe{
\item{School}{School (character)}
\item{State}{State (character)}
\item{W}{Wins (integer)}
\item{L}{Loss's (integer)}
\item{W.L}{Win Loss percentage (numeric)}
\item{SRS}{Simple Rating System (numeric)}
\item{SOS}{Strength of Schedule (numeric)}
\item{Points.Scored}{Points scored (integer)}
\item{Points.Allowed}{Points allowed (integer)}
\item{FG.}{Team field goal percentage (numeric)}
\item{X3P.}{Three point percentage (numeric)}
\item{FT.}{Free throw percentage (numeric)}
\item{Rebounds}{Number of rebounds (integer)}
\item{AST}{Number of assists (integer)}
\item{STL}{Number of steals (integer)}
\item{Blocks}{Number of blocks (integer)}
\item{Turn.Overs}{Number of turn overs (integer)}
\item{Fouls}{Number of fouls (integer)}
}

A data frame with 106 rows and 17 columns:
\describe{
\item{Player}{Players name (character)}
\item{Tm}{Players team (character)}
\item{Age}{Players Age (integer)}
\item{Pos}{Players position (character)}
\item{G}{Number of games (integer)}
\item{GS}{Number of games starting (integer)}
\item{Wins}{Number of wins (integer)}
\item{Cmp}{Number of completions (integer)}
\item{Att}{Number of throwing attempts (integer)}
\item{Cmp.}{Completion percentage (numeric)}
\item{Yds}{Number of yards thrown (integer)}
\item{TD}{Number of touchdowns (integer)}
\item{Int}{Number of interceptions thrown (integer)}
\item{Y.A}{Yards per Attempt (numeric)}
\item{Y.G}{Yards per Game (numeric)}
\item{Rate}{Passer rating (numeric)}
\item{QBR}{Total Quarterback Rating (numeric)}
}
}
\source{
Quarterback data

Quarterback data
}
\usage{
football

football
}
\description{
Ncaa basketball data

Football/quarterback data
}
\keyword{datasets}
