% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_data.R
\docType{data}
\name{housing_data}
\alias{housing_data}
\title{Housing data}
\format{
A data frame with 98 rows and 6 columns:
\describe{
\item{price}{Home price (numeric)}
\item{bed}{Number of bedrooms (integer)}
\item{bath}{Number of bathrooms (numeric)}
\item{area}{Square footage (integer)}
\item{year_built}{Date house was built (integer)}
\item{lot}{lot size (numeric)}
}
}
\source{
housing data
}
\usage{
housing_data
}
\description{
Housing data
}
\keyword{datasets}
