% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_data.R
\docType{data}
\name{income_data}
\alias{income_data}
\title{Income data}
\format{
A data frame with 20 rows and 5 columns:
\describe{
\item{ID}{ID (integer)}
\item{Ages}{age (integer)}
\item{Years_til_Retirement.65}{Years until retirement at 65 (integer)}
\item{Salary}{Salary (integer)}
\item{Birth_weight}{Birth weight (integer)}
}
}
\source{
Synthetic Data
}
\usage{
income_data
}
\description{
Dataset containing basic demographic and financial information for 20 individuals.
}
\keyword{datasets}
