% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_stats.R
\name{position_stats}
\alias{position_stats}
\title{Computes Position Statics, Quintiles and Quartiles}
\usage{
position_stats(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A list with two elements:
\describe{
\item{quint}{Numeric vector of quintiles (0\%, 20\%, 40\%, ..., 100\%)}
\item{quart}{Numeric vector of quartiles (0\%, 25\%, 50\%, 75\%, 100\%)}
}
}
\description{
Calculates the quintiles and quartiles of a numeric vector using the
'quintile()' function. NA's are removed
}
