% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time.R
\docType{data}
\name{reaction_time}
\alias{reaction_time}
\title{reaction data}
\format{
A data frame with 100 rows and 6 columns:
\describe{
\item{person}{Person id (integer)}
\item{color}{color (character)}
\item{left}{left (numeric)}
\item{right}{right (numeric)}
\item{age}{Person age (numeric)}
\item{Gender}{Person gender (character)}
}
}
\source{
reaction data
}
\usage{
reaction_time
}
\description{
reaction data
}
\keyword{datasets}
