% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running.R
\docType{data}
\name{running}
\alias{running}
\title{Running data}
\format{
A data frame with 16924 rows and 9 columns:
\describe{
\item{place}{running place (integer)}
\item{time}{running time (numeric)}
\item{pace}{running pace (numeric)}
\item{age}{Runners age (integer)}
\item{gender}{Runners gender (character)}
\item{location}{Location (character)}
\item{state}{State (character)}
\item{div_place}{Runners placement in given division (integer)}
\item{div_tot}{Total number of runners in the given division (integer)}
}
}
\source{
running data
}
\usage{
running
}
\description{
Running data
}
\keyword{datasets}
