% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Compute Sample Skewness}
\usage{
skew(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the skewness of the distribution.
}
\description{
Calculates the skewness of a numeric vector.
A positive value indicates right skew (long right tail), while a negative value
indicates left skew (long left tail). A zero value represents symmetry
}
\examples{
# Skew of Sepal Lengths in iris
data("iris")
skew(iris$Sepal.Length)


}
