% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer.R
\docType{data}
\name{soccer}
\alias{soccer}
\title{Historic soccer data}
\format{
A data frame with 13750 rows and 6 columns:
\describe{
\item{X}{numbered list (integer)}
\item{date}{Date of match (character)}
\item{home_team}{Home team name (character)}
\item{away_team}{Away team name (character)}
\item{home_score}{Home teams goal count (integer)}
\item{away_score}{Away teams goal count (integer)}
}
}
\source{
historic soccer data
}
\usage{
soccer
}
\description{
Historic soccer data
}
\keyword{datasets}
