% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_scores.R
\docType{data}
\name{course_scores}
\alias{course_scores}
\title{Course scores data}
\format{
A data frame with 200 rows and 10 columns:
\describe{
\item{student}{Student ID (integer)}
\item{type}{Grade type (character)}
\item{Freshman_English}{Freshman English Score/letter grade (character)}
\item{Freshman_Math}{Freshman Math Score/letter grade (character)}
\item{Sophomore_English}{Sophomore English Score/letter grade (character)}
\item{Sophomore_Math}{Sophomore Math Score/letter grade (character)}
\item{Junior_English}{Junior English Score/letter grade (character)}
\item{Junior_Math}{Junior Math Score/letter grade (character)}
\item{Senior_English}{Senior English Score/letter grade (character)}
\item{Senior_Math}{Senior Math Score/letter grade (character)}
}
}
\source{
course scores
}
\usage{
course_scores
}
\description{
Course scores data
}
\keyword{datasets}
