% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_210_census.R
\docType{data}
\name{data_210_census}
\alias{data_210_census}
\title{Synthetic Census dataset
A synthetic dataset containing demographic and socioeconomic information for
1,000 individuals.}
\format{
A data frame with 1000 rows and 5 columns:
\describe{
\item{age}{Persons Age (integer)}
\item{Gender}{Persons Gender (character)}
\item{degree}{Persons level of education (character)}
\item{salary}{Persons Yearly Salary (integer)}
\item{height}{Persons Height in inches (integer)}
}
}
\source{
Synthetic Data
}
\usage{
data_210_census
}
\description{
Synthetic Census dataset
A synthetic dataset containing demographic and socioeconomic information for
1,000 individuals.
}
\keyword{datasets}
