% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam_data.R
\docType{data}
\name{exam_data}
\alias{exam_data}
\title{Mini Project exam data}
\format{
A data frame with 1000 rows and 8 columns:
\describe{
\item{gender}{Students gender (character)}
\item{race.ethnicity}{Students race/ethnicity (character)}
\item{parental.level.of.education}{Parents level of education (character)}
\item{lunch}{Students lunch plan (character)}
\item{test.preparation.course}{Student test prep level (character)}
\item{math.score}{Students math score (integer)}
\item{reading.score}{Students reading score (integer)}
\item{writing.score}{Students writing score (integer)}
}
}
\source{
Mini project exam data
}
\usage{
exam_data
}
\description{
Mini Project exam data
}
\keyword{datasets}
