% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt.R
\name{kurt}
\alias{kurt}
\title{Compute Sample Kurtosis}
\usage{
kurt(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A single numeric value representing the kurtosis
}
\description{
Calculates the kurtosis of a numeric vector.
A value near 0 suggests normal kurtosis (mesokurtic),
positive values indicate heavier tails (leptokurtic), and negative
values indicate lighter tails (platykurtic).
}
