% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger_data.R
\docType{data}
\name{ledger_data}
\alias{ledger_data}
\title{Ledger data}
\format{
A data frame with 4 rows and 104 columns:
\describe{
\item{color}{colors (character)}
\item{type}{age (integer)}
\item{Jan_08}{Price on date (numeric)}
\item{Jan_15}{Price on date (numeric)}
\item{Jan_16}{Price on date (numeric)}
\item{Jan_31}{Price on date (numeric)}
\item{Feb_02}{Price on date (numeric)}
\item{Feb_03}{Price on date (numeric)}
\item{Feb_04}{Price on date (numeric)}
\item{Feb_14}{Price on date (numeric)}
\item{Feb_20}{Price on date (numeric)}
\item{Feb_22}{Price on date (numeric)}
\item{Feb_25}{Price on date (numeric)}
\item{Feb_27}{Price on date (numeric)}
\item{Feb_28}{Price on date (numeric)}
\item{Mar_01}{Price on date (numeric)}
\item{Mar_05}{Price on date (numeric)}
\item{Mar_09}{Price on date (numeric)}
\item{Mar_12}{Price on date (numeric)}
\item{Mar_16}{Price on date (numeric)}
\item{Mar_20}{Price on date (numeric)}
\item{Mar_21}{Price on date (numeric)}
\item{Mar_22}{Price on date (numeric)}
\item{Mar_24}{Price on date (numeric)}
\item{Mar_27}{Price on date (numeric)}
\item{Mar_28}{Price on date (numeric)}
\item{Mar_31}{Price on date (numeric)}
\item{Apr_06}{Price on date (numeric)}
\item{Apr_08}{Price on date (numeric)}
\item{Apr_10}{Price on date (numeric)}
\item{Apr_18}{Price on date (numeric)}
\item{Apr_19}{Price on date (numeric)}
\item{Apr_24}{Price on date (numeric)}
\item{Apr_26}{Price on date (numeric)}
\item{Apr_29}{Price on date (numeric)}
\item{May_01}{Price on date (numeric)}
\item{May_04}{Price on date (numeric)}
\item{May_12}{Price on date (numeric)}
\item{May_17}{Price on date (numeric)}
\item{May_24}{Price on date (numeric)}
\item{May_25}{Price on date (numeric)}
\item{May_28}{Price on date (numeric)}
\item{Jun_01}{Price on date (numeric)}
\item{Jun_04}{Price on date (numeric)}
\item{Jun_11}{Price on date (numeric)}
\item{Jun_16}{Price on date (numeric)}
\item{Jun_25}{Price on date (numeric)}
\item{Jun_28}{Price on date (numeric)}
\item{Jul_03}{Price on date (numeric)}
\item{Jul_04}{Price on date (numeric)}
\item{Jul_08}{Price on date (numeric)}
\item{Jul_10}{Price on date (numeric)}
\item{Jul_11}{Price on date (numeric)}
\item{Jul_13}{Price on date (numeric)}
\item{Jul_18}{Price on date (numeric)}
\item{Jul_23}{Price on date (numeric)}
\item{Jul_25}{Price on date (numeric)}
\item{Aug_05}{Price on date (numeric)}
\item{Aug_12}{Price on date (numeric)}
\item{Aug_13}{Price on date (numeric)}
\item{Aug_24}{Price on date (numeric)}
\item{Aug_26}{Price on date (numeric)}
\item{Sep_02}{Price on date (numeric)}
\item{Sep_06}{Price on date (numeric)}
\item{Sep_07}{Price on date (numeric)}
\item{Sep_08}{Price on date (numeric)}
\item{Sep_16}{Price on date (numeric)}
\item{Sep_21}{Price on date (numeric)}
\item{Sep_22}{Price on date (numeric)}
\item{Sep_23}{Price on date (numeric)}
\item{Sep_27}{Price on date (numeric)}
\item{Oct_07}{Price on date (numeric)}
\item{Oct_09}{Price on date (numeric)}
\item{Oct_10}{Price on date (numeric)}
\item{Oct_15}{Price on date (numeric)}
\item{Oct_16}{Price on date (numeric)}
\item{Oct_17}{Price on date (numeric)}
\item{Oct_19}{Price on date (numeric)}
\item{Oct_20}{Price on date (numeric)}
\item{Oct_21}{Price on date (numeric)}
\item{Oct_22}{Price on date (numeric)}
\item{Oct_29}{Price on date (numeric)}
\item{Oct_30}{Price on date (numeric)}
\item{Oct_31}{Price on date (numeric)}
\item{Nov_03}{Price on date (numeric)}
\item{Nov_04}{Price on date (numeric)}
\item{Nov_12}{Price on date (numeric)}
\item{Nov_13}{Price on date (numeric)}
\item{Nov_14}{Price on date (numeric)}
\item{Nov_16}{Price on date (numeric)}
\item{Nov_18}{Price on date (numeric)}
\item{Nov_23}{Price on date (numeric)}
\item{Nov_24}{Price on date (numeric)}
\item{Dec_02}{Price on date (numeric)}
\item{Dec_03}{Price on date (numeric)}
\item{Dec_06}{Price on date (numeric)}
\item{Dec_11}{Price on date (numeric)}
\item{Dec_12}{Price on date (numeric)}
\item{Dec_13}{Price on date (numeric)}
\item{Dec_16}{Price on date (numeric)}
\item{Dec_17}{Price on date (numeric)}
\item{Dec_18}{Price on date (numeric)}
\item{Dec_19}{Price on date (numeric)}
\item{Dec_26}{Price on date (numeric)}
}
}
\source{
Synthetic Data
}
\usage{
ledger_data
}
\description{
Dataset mimicking a ledger showing the price an item was bought and sold for,
the date it occurred, and the color of the product.
}
\keyword{datasets}
