% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger_data.R
\docType{data}
\name{ledger_data}
\alias{ledger_data}
\title{Ledger data}
\format{
A data frame with 4 rows and 104 columns:
\describe{
\item{color}{colors (character)}
\item{type}{age (integer)}
}
}
\source{
Synthetic Data
}
\usage{
ledger_data
}
\description{
Dataset mimicking a ledger showing the price an item was bought and sold for,
the date it occurred, and the color of the product.
}
\keyword{datasets}
