% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_time.R
\docType{data}
\name{reaction_time}
\alias{reaction_time}
\title{Reaction Data}
\format{
A data frame with 100 rows and 6 columns:
\describe{
\item{person}{Person id (integer)}
\item{color}{color (character)}
\item{left}{left (numeric)}
\item{right}{right (numeric)}
\item{age}{Person age (numeric)}
\item{Gender}{Person gender (character)}
}
}
\source{
Synthetic Data
}
\usage{
reaction_time
}
\description{
This dataset contains synthetic reaction time measurements for 100
individuals under different conditions.
}
\keyword{datasets}
